moved confirmation token logics into separate service
